+++ OutWit.Examples.ManagedWrapper/BenchmarkWrapper.h
@@ -4,7 +4,7 @@ namespace OutWit::Examples::ManagedWrapper
 	class __declspec(dllexport) BenchmarkWrapper
 	{
 	public:
-		static double RunInt(long iterations);
+		static long RunInt(long iterations);
 		static double RunDouble(long iterations);
 	};
 }
